requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.23.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pyyaml>=6.0.1
dvc>=3.40.1
fastapi>=0.110.0
uvicorn>=0.29.0
onnx>=1.15.0
onnxruntime>=1.18.0
requests>=2.31.0
tqdm>=4.66.0
